评论区引流器/-2/requirements.txt
# 评论区引流器项目依赖
# 核心依赖
selenium>=4.15.0
webdriver-manager>=4.0.1
pyautogui>=0.9.54
opencv-python>=4.8.1
numpy>=1.24.3
Pillow>=10.0.1

# 浏览器自动化
playwright>=1.41.0
python-dotenv>=1.0.0

# OCR和AI视觉
easyocr>=1.7.0
openai>=1.3.0
requests>=2.31.0

# 异步和并发
aiohttp>=3.9.0
aiofiles>=23.2.1

# 数据处理
pandas>=2.1.0
json5>=0.9.14

# 图像处理和计算机视觉
scipy>=1.11.0
scikit-image>=0.21.0
imageio>=2.31.0

# 验证码处理
captcha>=0.4

# 系统和工具
psutil>=5.9.0
colorlog>=6.7.0
tqdm>=4.66.0

# 开发和测试
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.7.0
flake8>=6.0.0
isort>=5.12.0

# 可选依赖（高级功能）
# tensorflow>=2.13.0  # 用于高级AI功能
# torch>=2.0.0       # 用于深度学习模型
# transformers>=4.30.0  # 用于NLP功能
